took get_riddle() out of hints.py, opened __main__
