+++ README.md
@@ -1,3 +1,18 @@
 # Python Package Exercise
 
-An exercise to create a Python package, build it, test it, distribute it, and use it. See [instructions](./instructions.md) for details.
+Replace the contents of the [README.md](./README.md) file with a beautifully-formatted Markdown file including
+
+- a plain-language **description** of your project, including:
+- a [badge](https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/adding-a-workflow-status-badge) at the top of the `README.md` file showing the result of the latest build/test workflow run.
+- a link to your package's page on the PyPI website.
+- how a developer who wants to import your project into their own code can do so - include documentation and code examples for all functions in your package and a link to an example Python program that uses each of them.
+- how a developer who wants to contribute to your project can set up the virtual environment, install dependencies, and build and test your package for themselves.
+- the names of all teammates as links to their GitHub profiles in the `README.md` file.
+- instructions for how to configure and run all parts of your project for any developer on any platform - these instructions must work!
+- instructions for how to set up any environment variables and import any starter data into the database, as necessary, for the system to operate correctly when run.
+- if there are any "secret" configuration files, such as `.env` or similar files, that are not included in the version control repository, examples of these files, such as `env.example`, with dummy data must be included in the repository and exact instructions for how to create the proper configuration files and what their contents should be must be supplied to the course admins by the due date.
+
+
+## Plain-language Description 
+
+- PyRiddles is a simple and fun python package for those who want to enjoy a few riddles! Users of the package would make use of it by accessing four main functions of
